requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
beautifulsoup4==4.12.3
bitarray==2.9.2
blinker==1.8.2
cachetools==5.3.3
cassandra-driver==3.29.1
cassio==0.1.8
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colbert-ai==0.2.19
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.19.2
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.1
Faker==25.8.0
favicon==0.7.0
filelock==3.14.0
Flask==3.0.3
fonttools==4.53.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.3.1
geomet==0.2.1.post1
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.64.1
grpcio-health-checking==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
humanize==4.9.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.4.5
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.76
llama-index-core==0.10.44
llamaindex-py-client==0.1.19
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
mmh3==4.1.0
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
openai==1.33.0
orjson==3.10.4
packaging==23.2
pandas==2.2.2
pillow==10.3.0
pinecone-client==4.1.1
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
plotly==5.22.0
prometheus_client==0.20.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==14.0.1
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.8.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
ragstack-ai-colbert==1.0.5
ragulate==0.0.10
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.35.0
streamlit-aggrid==0.3.4
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-pills==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12.1
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.41.2
trulens_eval==0.31.0
types-requests==2.32.0.20240602
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
validators==0.28.3
Werkzeug==3.0.3
wget==3.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
